sentinel_platform/sentinel_gateway/requirements.txt
fastapi
uvicorn
websockets
nats-py
pydantic-settings
python-jose
prometheus-fastapi-instrumentator